pyproject.toml
[tool.poetry]
name = "ai-chat-streamlit"
version = "0.1.0"
description = ""
authors = ["Bo Chen <bochen0909@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
datasets = "^2.19.1"
tqdm = "^4.66.4"
pandas = "^2.2.2"
openai = "^1.30.1"
streamlit = "^1.34.0"
langchain-community = "^0.0.38"
boto3 = "^1.34.105"
langchain-aws = "^0.1.3"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
